RPS Bonus: Changed Prompt slightly
Originally I was going to have the print_prompt method in Prompt be a private instance method. Problem is I have a need to print a user input prompt when getting a name and this doesn't exactly fall under prompts purpose. I decided to make print_prompt a class method of Prompt so that I can call it from the Human class.